FIX entrypoint.sh preventing regular and legitimate use-cases (#495)
This commit fixes the entrypoint.sh script in order to comply to the user provided
configurations. If they provided AIRFLOW__ environment variables, the entrypoint.sh
*must not* override them.

It also adds a new PostgreSQL configuration variable POSTGRES_EXTRAS in order to enable
TLS encryption. See the Airflow documentation "PostgreSQL Connection" about this.

It also adds two new Redis configuration variables REDIS_PROTO which was previously hardcoded
to redis:// (in order to prepare the possible use of rediss://), and REDIS_DBNUM which was
previously hardcoded to 1 (that prevents any use of a mutual broker between Airflow instances).

It also adds the corresponding documentation about PostgreSQL and Redis variables.